requirements.txt
numpy==1.19.2
tensorflow==2.4.1
Pillow==8.0.1
flask==1.1.2
flask_cors==3.0.10
requests==2.24.0

